+++ src/com/esotericsoftware/kryo/serializers/FieldSerializerGenericsUtil.java
@@ -33,8 +33,6 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
-import com.esotericsoftware.kryo.Generics;
-import com.esotericsoftware.kryo.GenericsResolver;
 import com.esotericsoftware.kryo.Kryo;
 import com.esotericsoftware.kryo.serializers.FieldSerializer.CachedField;
